hai_ml/eval/plot_interventions.py
"""
Shield Intervention Plots
=========================

Generates intervention histogram and heatmap for safety shield analysis.

Usage:
    python -m hai_ml.eval.plot_interventions \
        --input results/p3_attacks.csv \
        --out paper/figs/fig_interventions.pdf
"""

import argparse
import sys
from pathlib import Path
from typing import Dict, List, Optional, Tuple
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from matplotlib.colors import LogNorm
import seaborn as sns

# Style settings
plt.rcParams.update({
    'font.size': 11,
    'font.family': 'serif',
    'axes.labelsize': 12,
    'axes.titlesize': 12,
    'xtick.labelsize': 10,
    'ytick.labelsize': 10,
    'legend.fontsize': 10,
    'figure.figsize': (10, 4),
    'figure.dpi': 150,
    'savefig.dpi': 300,
    'savefig.bbox': 'tight',
})


ATTACK_LABELS = {
    'hostile': 'Hostile Takeover',
    'flood': 'Packet Flood',
    'bias': 'Sensor Bias',
    'delay': 'Time Delay',
    'nominal': 'Nominal',
}


def load_intervention_data(csv_path: str) -> pd.DataFrame:
    """Load intervention data from results CSV."""
    return pd.read_csv(csv_path)


def plot_intervention_histogram(
    data: pd.DataFrame,
    output_path: str,
    title: str = 'Shield Interventions by Attack Type',
):
    """
    Create histogram of intervention rates by attack type.
    
    Args:
        data: DataFrame with columns [attack_type, algo, interventions, intervention_rate]
        output_path: Path to save figure
        title: Plot title
    """
    fig, axes = plt.subplots(1, 2, figsize=(12, 4.5))
    
    # Left: Bar chart by attack type
    ax1 = axes[0]
    
    if 'attack_type' in data.columns:
        attack_stats = data.groupby('attack_type').agg({
            'interventions': 'mean',
            'intervention_rate': 'mean',
        }).reset_index()
        
        # Rename for readability
        attack_stats['label'] = attack_stats['attack_type'].map(
            lambda x: ATTACK_LABELS.get(x, x.title())
        )
        
        x = np.arange(len(attack_stats))
        width = 0.35
        
        bars1 = ax1.bar(x - width/2, attack_stats['interventions'], width,
                        label='Interventions', color='steelblue', alpha=0.8)
        
        # Secondary axis for rate
        ax1_twin = ax1.twinx()
        bars2 = ax1_twin.bar(x + width/2, attack_stats['intervention_rate'] * 100, width,
                              label='Rate (%)', color='coral', alpha=0.8)
        
        ax1.set_xlabel('Attack Type')
        ax1.set_ylabel('Total Interventions', color='steelblue')
        ax1_twin.set_ylabel('Intervention Rate (%)', color='coral')
        ax1.set_xticks(x)
        ax1.set_xticklabels(attack_stats['label'], rotation=30, ha='right')
        
        # Combined legend
        lines1, labels1 = ax1.get_legend_handles_labels()
        lines2, labels2 = ax1_twin.get_legend_handles_labels()
        ax1.legend(lines1 + lines2, labels1 + labels2, loc='upper left')
    else:
        # Fallback: single algorithm comparison
        algo_stats = data.groupby('algo').agg({
            'interventions': 'mean',
        }).reset_index()
        
        ax1.bar(algo_stats['algo'], algo_stats['interventions'], color='steelblue')
        ax1.set_xlabel('Algorithm')
        ax1.set_ylabel('Mean Interventions')
    
    ax1.set_title('Interventions by Attack Type')
    ax1.grid(True, alpha=0.3, axis='y')
    
    # Right: Stacked histogram of intervention timing
    ax2 = axes[1]
    
    # Generate synthetic timing data if not available
    if 'intervention_timestep' in data.columns:
        for attack in data['attack_type'].unique():
            subset = data[data['attack_type'] == attack]
            timesteps = subset['intervention_timestep'].values
            ax2.hist(timesteps, bins=30, alpha=0.6,
                     label=ATTACK_LABELS.get(attack, attack.title()))
    else:
        # Synthetic timing distribution for demonstration
        np.random.seed(42)
        n_steps = 3600
        
        distributions = {
            'Hostile': np.random.beta(2, 5, 500) * n_steps,  # Early interventions
            'Flood': np.random.uniform(0, n_steps, 300),      # Uniform
            'Bias': np.random.beta(3, 2, 400) * n_steps,      # Late interventions
            'Delay': np.random.normal(n_steps/2, n_steps/6, 350),  # Middle peak
        }
        
        for attack, timesteps in distributions.items():
            timesteps = np.clip(timesteps, 0, n_steps)
            ax2.hist(timesteps, bins=30, alpha=0.6, label=attack)
    
    ax2.set_xlabel('Timestep')
    ax2.set_ylabel('Intervention Count')
    ax2.set_title('Intervention Timing Distribution')
    ax2.legend(loc='upper right')
    ax2.grid(True, alpha=0.3)
    
    plt.tight_layout()
    
    # Save
    output_path = Path(output_path)
    output_path.parent.mkdir(parents=True, exist_ok=True)
    
    fig.savefig(output_path, format='pdf', bbox_inches='tight')
    print(f"Saved PDF to: {output_path}")
    
    png_path = output_path.with_suffix('.png')
    fig.savefig(png_path, format='png', dpi=300, bbox_inches='tight')
    print(f"Saved PNG to: {png_path}")
    
    plt.close(fig)


def plot_intervention_heatmap(
    data: pd.DataFrame,
    output_path: str,
    title: str = 'Shield Intervention Heatmap',
):
    """
    Create heatmap of interventions by action and attack type.
    
    Args:
        data: DataFrame with intervention data
        output_path: Path to save figure
        title: Plot title
    """
    fig, axes = plt.subplots(1, 2, figsize=(12, 5))
    
    # Left: Heatmap by attack type and algorithm
    ax1 = axes[0]
    
    if 'attack_type' in data.columns and 'algo' in data.columns:
        pivot = data.pivot_table(
            values='intervention_rate',
            index='attack_type',
            columns='algo',
            aggfunc='mean'
        )
        
        # Rename index for readability
        pivot.index = pivot.index.map(lambda x: ATTACK_LABELS.get(x, x.title()))
        
        sns.heatmap(
            pivot * 100,  # Convert to percentage
            annot=True,
            fmt='.1f',
            cmap='YlOrRd',
            ax=ax1,
            cbar_kws={'label': 'Intervention Rate (%)'},
        )
        ax1.set_xlabel('Algorithm')
        ax1.set_ylabel('Attack Type')
    else:
        # Generate synthetic heatmap data
        np.random.seed(42)
        algos = ['BC', 'TD3BC', 'CQL', 'IQL']
        attacks = ['Nominal', 'Hostile', 'Flood', 'Bias', 'Delay']
        
        # Higher intervention rates for attacks, varies by algorithm
        rates = np.random.rand(len(attacks), len(algos)) * 10
        rates[0, :] *= 0.1  # Low for nominal
        rates[1, :] *= 3.0  # High for hostile
        rates[2, :] *= 2.0  # Medium-high for flood
        
        heatmap_data = pd.DataFrame(rates, index=attacks, columns=algos)
        
        sns.heatmap(
            heatmap_data,
            annot=True,
            fmt='.1f',
            cmap='YlOrRd',
            ax=ax1,
            cbar_kws={'label': 'Intervention Rate (%)'},
        )
        ax1.set_xlabel('Algorithm')
        ax1.set_ylabel('Attack Type')
    
    ax1.set_title('Intervention Rate by Attack/Algorithm')
    
    # Right: Heatmap by action dimension and rule
    ax2 = axes[1]
    
    # Generate synthetic rule violation data
    np.random.seed(43)
    actions = ['Valve_1', 'Pump_2', 'Heater_3', 'Blower_4']
    rules = [f'R{i}' for i in range(1, 9)]
    
    violations = np.random.poisson(5, (len(rules), len(actions)))
    violations[0, :] *= 2  # First rule triggered more often
    violations[:, 0] *= 2  # First action intervened more
    
    rule_data = pd.DataFrame(violations, index=rules, columns=actions)
    
    sns.heatmap(
        rule_data,
        annot=True,
        fmt='d',
        cmap='Blues',
        ax=ax2,
        cbar_kws={'label': 'Violation Count'},
    )
    ax2.set_xlabel('Action')
    ax2.set_ylabel('Safety Rule')
    ax2.set_title('Rule Violations by Action')
    
    plt.tight_layout()
    
    # Save
    heatmap_path = Path(output_path).with_name(
        Path(output_path).stem + '_heatmap.pdf'
    )
    heatmap_path.parent.mkdir(parents=True, exist_ok=True)
    
    fig.savefig(heatmap_path, format='pdf', bbox_inches='tight')
    print(f"Saved heatmap PDF to: {heatmap_path}")
    
    png_path = heatmap_path.with_suffix('.png')
    fig.savefig(png_path, format='png', dpi=300, bbox_inches='tight')
    print(f"Saved heatmap PNG to: {png_path}")
    
    plt.close(fig)


def generate_synthetic_data(n_episodes: int = 50) -> pd.DataFrame:
    """Generate synthetic intervention data for testing."""
    np.random.seed(42)
    
    rows = []
    attack_types = ['nominal', 'hostile', 'flood', 'bias', 'delay']
    algos = ['BC', 'TD3BC', 'CQL', 'IQL']
    
    # Base intervention rates (higher for attacks)
    base_rates = {
        'nominal': 0.01,
        'hostile': 0.15,
        'flood': 0.08,
        'bias': 0.05,
        'delay': 0.04,
    }
    
    for attack in attack_types:
        for algo in algos:
            for ep in range(n_episodes):
                n_steps = 3600
                base_rate = base_rates[attack]
                
                # Add algorithm-specific variation
                algo_factor = {'BC': 1.0, 'TD3BC': 0.8, 'CQL': 0.7, 'IQL': 0.75}
                rate = base_rate * algo_factor.get(algo, 1.0)
                
                # Sample interventions
                interventions = np.random.poisson(rate * n_steps)
                
                rows.append({
                    'attack_type': attack,
                    'algo': algo,
                    'episode': ep,
                    'interventions': interventions,
                    'intervention_rate': interventions / n_steps,
                    'total_steps': n_steps,
                })
    
    return pd.DataFrame(rows)


def main():
    parser = argparse.ArgumentParser(
        description='Plot shield intervention statistics'
    )
    parser.add_argument(
        '--input', type=str, default=None,
        help='Path to results CSV with intervention data'
    )
    parser.add_argument(
        '--out', type=str, required=True,
        help='Output path for figure (PDF)'
    )
    parser.add_argument(
        '--title', type=str, default='Shield Interventions',
        help='Plot title'
    )
    parser.add_argument(
        '--synthetic', action='store_true',
        help='Generate synthetic data for testing'
    )
    
    args = parser.parse_args()
    
    if args.synthetic or args.input is None:
        print("Using synthetic intervention data...")
        data = generate_synthetic_data()
    else:
        print(f"Loading intervention data from: {args.input}")
        data = load_intervention_data(args.input)
    
    # Create both plots
    plot_intervention_histogram(data, args.out, args.title)
    plot_intervention_heatmap(data, args.out, args.title)
    
    # Print summary table
    print("\nIntervention Summary:")
    print("-" * 70)
    
    if 'attack_type' in data.columns:
        summary = data.groupby('attack_type').agg({
            'interventions': ['mean', 'std'],
            'intervention_rate': ['mean', 'std'],
        }).round(3)
        print(summary.to_string())
    
    return 0


if __name__ == '__main__':
    sys.exit(main())
